requirements.txt
annotated-types==0.6.0
Babel==2.13.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
docutils==0.20.1
docxcompose==1.4.0
# docxtpl==0.16.7 # - текущая версия библиотеки сломана. Я внес исправления и добавил в проект рабочую версию
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.17.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.61.0
gspread==5.8.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
pockets==0.9.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-docx==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
socks==0
Sphinx==1.2.3
sphinxcontrib-napoleon==0.7
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
win-unicode-console==0.5
